feat: Add migration tracking for ESLint rules
This commit introduces a `migration-progress.md` file to track the testing status of all ESLint rules during the migration to ESLint v9.

It also refactors the existing tests for custom rules into individual, dedicated test files. The tests for the built-in rules have been marked as pending to be addressed in a future commit.
