+++ eslint.config.js
@@ -93,6 +93,8 @@ module.exports = [
       "no-constant-condition": "warn",
       "no-control-regex": "warn",
       "no-debugger": "warn",
+      "no-delete-var": "warn",
+      "no-dupe-args": "warn",
       "no-dupe-class-members": "warn",
       "no-dupe-keys": "warn",
       "no-duplicate-case": "warn",
@@ -112,8 +114,9 @@ module.exports = [
       "no-implied-eval": "warn",
       "no-invalid-regexp": "warn",
       "no-irregular-whitespace": "warn",
+      "no-iterator": "warn",
       "no-label-var": "warn",
-      "no-labels": ["warn", { "allowLoop": true, "allowSwitch": false }],
+      "no-labels": ["warn", { "allowLoop": false, "allowSwitch": false }],
       "no-lone-blocks": "warn",
       "no-loop-func": "warn",
       "no-mixed-spaces-and-tabs": "warn",
@@ -148,12 +151,15 @@ module.exports = [
         },
       ],
       "no-multi-str": "warn",
+      "no-native-reassign": "warn",
+      "no-negated-in-lhs": "warn",
       "no-new-func": "warn",
       "no-new-object": "warn",
       "no-new-symbol": "warn",
       "no-new-wrappers": "warn",
       "no-obj-calls": "warn",
       "no-octal": "warn",
+      "no-octal-escape": "warn",
       "no-redeclare": "warn",
       "no-regex-spaces": "warn",
       "no-restricted-globals": ["warn", "addEventListener", "blur", "close", "closed", "confirm", "defaultStatus", "defaultstatus", "event", "external", "find", "focus", "frameElement", "frames", "history", "innerHeight", "innerWidth", "length", "location", "locationbar", "menubar", "moveBy", "moveTo", "name", "onblur", "onwarn", "onfocus", "onload", "onresize", "onunload", "open", "opener", "opera", "outerHeight", "outerWidth", "pageXOffset", "pageYOffset", "parent", "print", "removeEventListener", "resizeBy", "resizeTo", "screen", "screenLeft", "screenTop", "screenX", "screenY", "scroll", "scrollbars", "scrollBy", "scrollTo", "scrollX", "scrollY", "self", "status", "statusbar", "stop", "toolbar", "top"],
